type fixes for cross-compiling (e.g. with RIOT)

+++ ccn-lite/src/ccnl-core/src/ccnl-prefix.c
@@ -185,7 +185,7 @@ ccnl_prefix_addChunkNum(struct ccnl_prefix_s *prefix, uint32_t chunknum)
 {
     if (chunknum >= 0xff) {
       DEBUGMSG_CUTL(WARNING, "addChunkNum is only implemented for "
-               "chunknum smaller than 0xff (%d)\n", chunknum);
+               "chunknum smaller than 0xff (%lu)\n", (long unsigned) chunknum);
         return -1;
     }
 
@@ -517,15 +517,15 @@ ccnl_prefix_cmp(struct ccnl_prefix_s *pfx, unsigned char *md,
         clen = i < pfx->compcnt ? pfx->complen[i] : 32; // SHA256_DIGEST_LEN
         if (clen != nam->complen[i] || memcmp(comp, nam->comp[i], nam->complen[i])) {
             rc = mode == CMP_EXACT ? -1 : (int32_t) i;
-            DEBUGMSG(VERBOSE, "component mismatch: %i\n", i);
+            DEBUGMSG(VERBOSE, "component mismatch: %lu\n", (long unsigned)i);
             goto done;
         }
     }
     // FIXME: we must also inspect chunknum here!
     rc = (mode == CMP_EXACT) ? 0 : (int32_t) i;
     done:
-    DEBUGMSG(TRACE, "  cmp result: pfxlen=%d cmplen=%d namlen=%d matchlen=%d\n",
-             pfx->compcnt, plen, nam->compcnt, rc);
+    DEBUGMSG(TRACE, "  cmp result: pfxlen=%lu cmplen=%lu namlen=%lu matchlen=%ld\n",
+             (long unsigned)pfx->compcnt, (long unsigned)plen, (long unsigned)nam->compcnt, (long) rc);
     return rc;
 }
 
@@ -794,7 +794,7 @@ ccnl_prefix_debug_info(struct ccnl_prefix_s *p) {
     }
     len += result;
 
-    result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "suite:%i, ", p->suite);
+    result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "suite:%lu, ", (long unsigned)p->suite);
     if (!(result > -1 && (unsigned)result < (CCNL_MAX_PACKET_SIZE - len))) {
         DEBUGMSG(ERROR, "Could not print prefix, since out of allocated memory");
         ccnl_free(buf);
@@ -803,7 +803,7 @@ ccnl_prefix_debug_info(struct ccnl_prefix_s *p) {
     len += result;
 
 #ifdef USE_NFN
-    result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "nfnflags:%i (", p->nfnflags);
+    result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "nfnflags:%lu (", (long unsigned)p->nfnflags);
     if (!(result > -1 && (unsigned)result < (CCNL_MAX_PACKET_SIZE - len))) {
         DEBUGMSG(ERROR, "Could not print prefix, since out of allocated memory");
         ccnl_free(buf);
@@ -813,7 +813,7 @@ ccnl_prefix_debug_info(struct ccnl_prefix_s *p) {
     int flagcount = 6;
     char *flagnames[6]  = {"NFN", "?", "COMPUTE", "KEEPALIVE", "INTERMEDIATE", "REQUEST"};
     int needscomma = 0;
-    for (i = 0; i < flagcount; i++) {
+    for (i = 0;  i < (unsigned) flagcount; i++) {
         if ((p->nfnflags & (1<<i)) != 0) {
             if (needscomma) {
                 result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, ",");
@@ -844,7 +844,7 @@ ccnl_prefix_debug_info(struct ccnl_prefix_s *p) {
 
 #endif
 
-    result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "compcnt:%i ", p->compcnt);
+    result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "compcnt:%lu ",(long unsigned) p->compcnt);
     if (!(result > -1 && (unsigned)result < (CCNL_MAX_PACKET_SIZE - len))) {
         DEBUGMSG(ERROR, "Could not print prefix, since out of allocated memory");
         ccnl_free(buf);
@@ -893,7 +893,7 @@ ccnl_prefix_debug_info(struct ccnl_prefix_s *p) {
     }
     len += result;
     for (i = 0; i < p->compcnt; i++) {
-        result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "%.*s", (uint32_t)p->complen[i], p->comp[i]);
+        result = snprintf(buf + len, CCNL_MAX_PACKET_SIZE - len, "%.*s", (int)p->complen[i], p->comp[i]);
         if (!(result > -1 && (unsigned)result < (CCNL_MAX_PACKET_SIZE - len))) {
             DEBUGMSG(ERROR, "Could not print prefix, since out of allocated memory");
             ccnl_free(buf);
